all cascade test finish

+++ web/src/main/java/com/vege/service/impl/DiseaseServiceImpl.java
@@ -1,13 +1,16 @@
 package com.vege.service.impl;
 
 import com.vege.dao.DiseaseRepository;
+import com.vege.dao.VegeInfoRepository;
 import com.vege.model.Disease;
+import com.vege.model.VegeInfo;
 import com.vege.service.DiseaseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -19,10 +22,17 @@ public class DiseaseServiceImpl extends BaseService implements DiseaseService {
     @Autowired
     DiseaseRepository diseaseRepository;
 
+    @Autowired
+    VegeInfoRepository vegeInfoRepository;
+
     @Override
+    @Transactional
     public boolean add(Disease disease) {
         try {
-            diseaseRepository.save(disease);
+            diseaseRepository.saveAndFlush(disease);
+            VegeInfo vegeInfo = disease.getVegeInfo();
+            vegeInfo.getDiseases().add(disease);
+            vegeInfoRepository.save(vegeInfo);
             return true;
         }catch (Exception e){
             return false;
@@ -33,6 +43,12 @@ public class DiseaseServiceImpl extends BaseService implements DiseaseService {
     public boolean delete(String diseaseId) {
         try {
             Disease disease = diseaseRepository.findByDiseaseId(Integer.parseInt(diseaseId));
+            if(disease.getSymptoms().size()!=0 || disease.getCures().size()!=0){
+                return false;
+            }
+            VegeInfo vegeInfo = disease.getVegeInfo();
+            vegeInfo.getDiseases().remove(disease);
+            vegeInfoRepository.save(vegeInfo);
             diseaseRepository.delete(disease);
             return true;
         }catch (Exception e){
